Initial interactive nearest-neighbor map viewer.
This is a big commit. The bulk of it is in the spatial_map_viewer_app. This is a Shiny app that allows user to pick two phenotypes and a single field, and show nearest-neighbor relationships within the field.

There is no GUI for selection of the data sources for the viewer. The shiny_map_viewer function is a simple front-end to the app.

The actual visualization is done by nearest_neighbor_map, which creates a ggplot object showing the composite image and nearest neighbors.